services/classifier/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pyyaml==6.0.1

# LangChain core
langchain>=0.2.10,<0.3
langchain-community>=0.2.10,<0.3
langchain-huggingface>=0.0.3
langchain-qdrant>=0.1.4,<0.2

qdrant-client>=1.10.1,<2.0

# HF embedding stack
sentence-transformers==2.7.0
transformers==4.42.4
accelerate==0.33.0